+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+### v1.0.4 (2021-07-30)
+
+- fix `ali-oss` error
+- `esc` event (exit fullscreen) for universal editor
+- `upload image` for universal editor
+- markdown coper for image uploader
+
 ### v1.0.3 (2021-07-27)
 
 - replace `service/loading` to [`veact-use`](https://github.com/veactjs/veact-use).